Sean (#8)
* login implementation

* update signup and login encryption method, added friend list and friend detail

* working on deleting invitation

* done login, working on create group

* add create group query

* working on create group

* added invite querys

* add update profile and change password

* modify database for spliting expense between selected user and provide function for retrieving expense detail information

* add local ref mydb to every function

* updated functions

* fix login

* fix some user exists and log in sign up buggys

* added logging system

* fix output

* add leave group function

* fix leave group, added comments

* rename app

* add log file

* delete dash function

* added sql files

* add zip to gitignore

* add display userlist within a group when adding expense

* fix addInvite missing groupid

* addComment function

* defined a friendListDisplay function

+++ src/sqls.py
@@ -1,6 +1,9 @@
+# SJSU CMPE 226 Spring 2021 TEAM4
 import bcrypt
 import hashlib
 import mysql.connector
+import logging
+logging.basicConfig(filename='EZledger.log', filemode='a', format='%(asctime)s - %(message)s', level=logging.INFO)
 hostname = 'database-1.c1sqwxkauabb.us-west-1.rds.amazonaws.com'
 username = 'admin'
 password = '19950808'
@@ -44,6 +47,7 @@ def insertNewUser(email, name, password):
     mydb.commit()
 
     print(mycursor.rowcount, "record inserted.")
+    logging.info('New user inserted: {}, {}, {}'.format(email, name, password))
 
 
 def userLogin(email, password):
@@ -53,6 +57,8 @@ def userLogin(email, password):
     res = []
     for x in mycursor:
         res.append(x)
+
+    logging.info('User login: {} {}'.format(email, password))
 
     # if the password is correct, return the user information. Otherwise, return empty list
     if res and bcrypt.checkpw(password.encode('utf8'), res[0][2].encode('utf8')):
@@ -64,6 +70,7 @@ def userLogin(email, password):
 
 
 def returnAllUsers():
+    logging.info('Return all users')
     mydb = mysql.connector.connect(
         host=hostname,
         user=username,
@@ -78,6 +85,7 @@ def returnAllUsers():
     return res
 
 def returnAllGroups():
+    logging.info('Return all groups')
     mydb = mysql.connector.connect(
         host=hostname,
         user=username,
@@ -93,6 +101,7 @@ def returnAllGroups():
 
 
 def invitations(whoami):
+    logging.info('View invitations - {}'.format(whoami))
     ''' View active invitations from the invite table '''
     mydb = mysql.connector.connect(
         host=hostname,
@@ -127,6 +136,7 @@ def returnDebts(whoami):
 
 
 def groupList(whoami):
+    logging.info('groupList - {}'.format(whoami))
     ''' This function returns the groups that I am currently in '''
     mydb = mysql.connector.connect(
         host=hostname,
@@ -154,15 +164,27 @@ def leaveGroup(whoami, groupNo):
             database=database
         )
         mycursor = mydb.cursor()
+
+        sql = "SELECT sum(balance) FROM debt where user1={} and group_id={}".format(whoami, groupNo)
+        mycursor.execute(sql)
+
+        res = []
+        for x in mycursor:
+            res.append(x)
+
+        if res[0][0] != 0:
+            print("Balance not clear, settle up with other user before you leave!")
+            return;
+        
         mycursor.execute("DELETE FROM groups_users WHERE user_id = '{}' and group_id = '{}'".format(
             str(whoami), str(groupNo)))
-
-    except:
-        print('Something went wrong in leaveGroup')
-    finally:
         mydb.commit()
         print(mycursor.rowcount, "record(s) deleted")
         return
+
+    except:
+        print('Something went wrong in leaveGroup')
+        
 
 
 def addToGroup(whomToAdd, groupNo):
@@ -200,6 +222,7 @@ def deleteInvitation(whoami):
 # display all the friends and debt
 # Display all debts from a certain user across all groups
 def friendList(whoami):
+    logging.info('CALL FindFriends - {}'.format(whoami))
     sql = "CALL FindFriends({})".format(whoami)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -218,6 +241,7 @@ def friendList(whoami):
 
 # display detail for a certain friend, like how much is owed in which group
 def friendDetail(whoami, friend_id):
+    logging.info('CALL FriendDetail - {} {}'.format(whoami, friend_id))
     sql = "CALL FriendDetail({}, {})".format(whoami, friend_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -237,6 +261,7 @@ def friendDetail(whoami, friend_id):
 
 # Create group will create a group and add the user created it to the group
 def createGroup(group_name, user_id):
+    logging.info('CALL CreateGroup - {} {}'.format(group_name, user_id))
     sql = "CALL CreateGroup(\"{}\", {});".format(group_name, user_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -256,6 +281,7 @@ def createGroup(group_name, user_id):
 
 # This should return all the activity inside this group
 def groupActivity(group_id):
+    logging.info('groupActivity - {}'.format(group_id))
     sql = "select * from `groups` g join expense e on g.group_id = e.group_id where g.group_id = {} ORDER BY e.`time` DESC ".format(group_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -275,6 +301,7 @@ def groupActivity(group_id):
 # {'detail': [{'owedBy': 'user1', 'amount': 10.0}, {'owedBy': 'user2', 'amount': 10.0}, {'owedBy': 'user3', 'amount': 10.0}], 'name': 'test4', 'paidBy': 'charles', 'totalAmount': 40.0}
 # detail is how much each user owes the user paid this expense. other is detail of this expense
 def activityDetail(expense_id):
+    logging.info('activityDetail - {}'.format(expense_id))
     sql = "SELECT u1.name as paid_username, e.name, e.cost as total_amount, u2.name as owe_username, se.amount FROM sub_expense se JOIN expense e on se.expense_id=e.expense_id JOIN user u1 on e.user_id = u1.user_id JOIN user u2 on se.user2 = u2.user_id WHERE e.expense_id={}".format(expense_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -298,6 +325,7 @@ def activityDetail(expense_id):
 
 
 def settleBalance(user1, user2):
+    logging.info('settleBalance - {} {}'.format(user1, user2))
     sql = "UPDATE debt set balance = 0 where user1={} AND user2={} OR user1={} AND user2={}".format(user1, user2, user2, user1)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -321,6 +349,7 @@ def settleBalance(user1, user2):
 # sample usage: addInvite(user_list=["1@user.com", "2@user.com"], group_id=51)
 # if false, abort all add invite query.
 def addInvite(user_list, group_id):
+    logging.info('addInvite - {}'.format(user_list, group_id))
     val = []
     for user in user_list:
         val.append((group_id, user))
@@ -344,6 +373,7 @@ def addInvite(user_list, group_id):
 
 # accept invite, need user's email and the group_id user invited to
 def acceptInvite(group_id, email):
+    logging.info('CALL AcceptInvite - {} {}'.format(group_id, email))
     sql = "CALL AcceptInvite({}, \"{}\")".format(group_id, str(email))
     mydb = mysql.connector.connect(
       host=hostname,
@@ -363,6 +393,7 @@ def acceptInvite(group_id, email):
 
 # decline invite, just delete that invite
 def declineInvite(group_id, email):
+    logging.info('declineInvite - {} {}'.format(group_id, email))
     sql = "DELETE FROM group_invite WHERE group_id = {} and email = \'{}\'".format(group_id, email)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -381,6 +412,7 @@ def declineInvite(group_id, email):
 
 
 def updateProfile(user_id, new_profile):
+    logging.info('updateProfile - {} {}'.format(user_id, new_profile))
     sql = "UPDATE user SET `name` = '{}', `email` = '{}' WHERE `user_id`='{}'".format(new_profile['name'], new_profile['email'], user_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -399,6 +431,7 @@ def updateProfile(user_id, new_profile):
 
 
 def changePassword(user_id, orig_password, new_password):
+    logging.info('changePassword -  {} {} {}'.format(user_id, orig_password, new_password))
     sql = "select email, name, password from user where user_id = '{}'".format(user_id)
     mydb = mysql.connector.connect(
       host=hostname,
@@ -433,6 +466,7 @@ def changePassword(user_id, orig_password, new_password):
 
 
 def addExpense(paid_by, user_list, amount, group_id, name):
+    logging.info('addExpense - {} {} {} {}'.format(paid_by, user_list, amount, group_id, name ))
     sql = "INSERT INTO expense(user_id, group_id, `name`, cost) VALUES (%s, %s, %s, %s)"
     val = (paid_by, group_id, name, amount)
     mydb = mysql.connector.connect(
@@ -465,14 +499,80 @@ def addExpense(paid_by, user_list, amount, group_id, name):
     else:
         return False
 
+def addComment(expense_id, user_id, content):
+    mydb = mysql.connector.connect(
+        host=hostname,
+        user=username,
+        password=password,
+        database=database
+    )
+    mycursor = mydb.cursor()
+
+    sql = "INSERT INTO comment(expense_id, user_id, content) VALUES(%s, %s, %s)"
+    val = (expense_id, user_id, content)
+    mycursor.execute(sql, val)
+    if mycursor.rowcount > 0:
+        mydb.commit()
+        return True
+    else:
+        return False
+
+
+def getComment(expense_id):
+    mydb = mysql.connector.connect(
+        host=hostname,
+        user=username,
+        password=password,
+        database=database
+    )
+    mycursor = mydb.cursor()
+
+    sql = "SELECT u.name, c.content, time FROM comment c JOIN user u on c.user_id=u.user_id WHERE expense_id={}".format(expense_id)
+    mycursor.execute(sql)
+    
+    res = []
+    for x in mycursor:
+        res.append(x)
+
+    return res
+
+def getGroupUsers(group_id):
+    mydb = mysql.connector.connect(
+        host=hostname,
+        user=username,
+        password=password,
+        database=database
+    )
+    mycursor = mydb.cursor()
+
+    sql = "SELECT u.name, u.user_id from groups_users g join user u on u.user_id = g.user_id where group_id = {}".format(group_id)
+    mycursor.execute(sql)
+    
+    res = []
+    for x in mycursor:
+        res.append(x)
+
+    return res
+
+def returnAllExpenses():
+    mydb = mysql.connector.connect(
+        host=hostname,
+        user=username,
+        password=password,
+        database=database
+    )
+    mycursor = mydb.cursor()
+
+    sql = 'SELECT * FROM expense'
+    mycursor.execute(sql)
+    return [x for x in mycursor]
+
+
 
 if __name__ == '__main__':
 
-    x = input('Enter a list\n')
-    if x:
-        print(x)
-    else:
-        print('Nothing')
+    a = returnAllExpenses()
+    print(a)
